Fix startup with low speeds
The Nvidia driver on my machine uses a speed of 32% by default when manual fan control is activated. This script on the other hand assumed that the default was 0%. As a result, it did not try to change the speed and the fan kept running at 32% even though temperatures were low enough to switch it off. By setting the initial speed to 100% instead, the script is guaranteed to adjust the speed downwards.